Move, use ByteSizeFormatter
