+++ sources/Google.Solutions.IapDesktop.Application.Test/Google.Solutions.IapDesktop.Application.Test.csproj
@@ -135,7 +135,6 @@
     <Compile Include="Settings\TestRegistrySecureStringSetting.cs" />
     <Compile Include="Settings\TestRegistryStringSetting.cs" />
     <Compile Include="Host\TestCommandLineOptions.cs" />
-    <Compile Include="Util\TestByteSizeFormatter.cs" />
     <Compile Include="Services\SecureConnect\TestSecureConnectEnrollment.cs" />
     <Compile Include="Services\Integration\TestJobService.cs" />
     <Compile Include="ObjectModel\TestServiceRegistry.cs" />
