+++ sources/Google.Solutions.IapDesktop.Extensions.Shell/Views/TunnelsViewer/TunnelsWindow.cs
@@ -26,6 +26,7 @@ using Google.Solutions.IapDesktop.Application.Views;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.Tunnel;
 using Google.Solutions.Mvvm.Binding;
 using Google.Solutions.Mvvm.Controls;
+using Google.Solutions.Mvvm.Shell.Util;
 using System;
 using System.Runtime.InteropServices;
 using WeifenLuo.WinFormsUI.Docking;
