+++ sources/Google.Solutions.Mvvm.Test/Google.Solutions.Mvvm.Test.csproj
@@ -81,6 +81,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Shell\TestFileType.cs" />
     <Compile Include="Shell\TestFileTypeCache.cs" />
+    <Compile Include="Util\TestByteSizeFormatter.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
